requirements.txt
amqp==2.5.1
appdirs==1.4.3
# asgi-redis==1.4.3
asgiref~=3.0
async-timeout==3.0.1
bandit==1.6.2
billiard==3.6.1.0
boto==2.49.0
boto3==1.9.233
botocore==1.12.233
channels-redis==2.4.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.5
django-allauth==0.40.0
django-appconf==1.0.3
django-betterforms==1.2
django-celery-email==2.0.2
django-classy-tags==0.9.0
django-colorfield==0.1.15
django-contrib-comments==1.9.1
django-cors-headers==3.1.0
django-crum==0.7.3
django-debug-toolbar==2.0
django-extra-views==0.12.0
django-fcm==0.1.1
django-filter==2.2.0
django-model-utils==3.2.0
django-nose==1.4.6
django-notifications-hq==1.5.0
django-postgrespool==0.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-storages==1.7.2
django-taggit==1.1.0
django-taggit-templatetags2==1.6.1
django-versatileimagefield==1.10
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
docutils==0.15.2
drfdocs2==0.0.13
flake8==3.7.8
flake8-polyfill==1.0.2
gevent==1.4.0
gitdb2==2.0.5
GitPython==3.0.2
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonfield==2.0.2
kombu==4.6.4
mando==0.6.4
MarkupSafe==1.1.1
mccabe==0.6.1
model_mommy==1.6.0
msgpack-python==0.5.6
nose==1.3.7
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
pbr==5.4.3
Pillow<6.0.0,>=2.4.0
pinocchio==0.4.2
psycopg2==2.8.3
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
radon==4.0.0
raven==6.10.0
redis==3.2.0
requests==2.22.0
requests-oauthlib==1.2.0
selenium<4
s3transfer==0.2.1
simplejson==3.16.0
six==1.12.0
smmap2==2.0.5
slackclient==2.1.0
SQLAlchemy==1.3.8
sqlparse==0.3.0
stevedore==1.31.0
uritemplate==3.0.0
urllib3==1.25.5
vine==1.3.0
psycopg2-binary==2.8.3
django-select2==7.1.1
